[utils] extract receipt bounding box helpers
