+++ portfolio/components/ui/Figures/PhotoReceiptBoundingBox.tsx
@@ -18,6 +18,11 @@ import {
   findBoundaryLinesWithSkew,
   estimateReceiptPolygonFromLines,
 } from "../../../utils/receiptGeometry";
+import {
+  findHullExtentsRelativeToCentroid,
+  computeReceiptBoxFromHull,
+  findLineEdgesAtPrimaryExtremes,
+} from "../../../utils/receiptBoundingBox";
 
 // Define simple point and line-segment shapes
 const isDevelopment = process.env.NODE_ENV === "development";
@@ -31,205 +36,6 @@ type LineSegment = {
 };
 
 
-/**
- * Find hull extents relative to centroid (matching Python implementation)
- */
-const findHullExtentsRelativeToCentroid = (
-  hull: Point[],
-  centroid: Point
-): {
-  minX: number;
-  maxX: number;
-  minY: number;
-  maxY: number;
-  leftPoint: Point;
-  rightPoint: Point;
-  topPoint: Point;
-  bottomPoint: Point;
-} => {
-  let minX = Infinity,
-    maxX = -Infinity;
-  let minY = Infinity,
-    maxY = -Infinity;
-  let leftPoint = hull[0],
-    rightPoint = hull[0];
-  let topPoint = hull[0],
-    bottomPoint = hull[0];
-
-  hull.forEach((point) => {
-    const relX = point.x - centroid.x;
-    const relY = point.y - centroid.y;
-
-    if (relX < minX) {
-      minX = relX;
-      leftPoint = point;
-    }
-    if (relX > maxX) {
-      maxX = relX;
-      rightPoint = point;
-    }
-    if (relY < minY) {
-      minY = relY;
-      bottomPoint = point;
-    }
-    if (relY > maxY) {
-      maxY = relY;
-      topPoint = point;
-    }
-  });
-
-  return {
-    minX,
-    maxX,
-    minY,
-    maxY,
-    leftPoint,
-    rightPoint,
-    topPoint,
-    bottomPoint,
-  };
-};
-
-/**
- * Compute receipt box from hull using the same algorithm as Python
- */
-const computeReceiptBoxFromHull = (
-  hull: Point[],
-  centroid: Point,
-  avgAngle: number
-): Point[] => {
-  if (hull.length < 3) return [];
-
-  // Find extents of hull relative to centroid
-  let minX = Infinity,
-    maxX = -Infinity;
-  let minY = Infinity,
-    maxY = -Infinity;
-
-  // Rotate points by negative average angle to align with receipt orientation
-  const angleRad = (-avgAngle * Math.PI) / 180;
-  const cosA = Math.cos(angleRad);
-  const sinA = Math.sin(angleRad);
-
-  hull.forEach((point) => {
-    // Translate to centroid origin
-    const relX = point.x - centroid.x;
-    const relY = point.y - centroid.y;
-
-    // Rotate
-    const rotX = relX * cosA - relY * sinA;
-    const rotY = relX * sinA + relY * cosA;
-
-    minX = Math.min(minX, rotX);
-    maxX = Math.max(maxX, rotX);
-    minY = Math.min(minY, rotY);
-    maxY = Math.max(maxY, rotY);
-  });
-
-  // Create bounding box corners in rotated space
-  const corners = [
-    { x: minX, y: maxY }, // top-left
-    { x: maxX, y: maxY }, // top-right
-    { x: maxX, y: minY }, // bottom-right
-    { x: minX, y: minY }, // bottom-left
-  ];
-
-  // Rotate back and translate to world coordinates
-  const reverseAngleRad = -angleRad;
-  const cosRA = Math.cos(reverseAngleRad);
-  const sinRA = Math.sin(reverseAngleRad);
-
-  return corners.map((corner) => ({
-    x: corner.x * cosRA - corner.y * sinRA + centroid.x,
-    y: corner.x * sinRA + corner.y * cosRA + centroid.y,
-  }));
-};
-
-/**
- * Find the left and right boundary lines using perpendicular projection
- */
-const findLineEdgesAtPrimaryExtremes = (
-  lines: Line[],
-  hull: Point[],
-  centroid: Point,
-  avgAngle: number
-): {
-  leftEdge: Point[];
-  rightEdge: Point[];
-} => {
-  const angleRad = (avgAngle * Math.PI) / 180;
-  const primaryAxisAngle = angleRad;
-
-  const secondaryAxisAngle = primaryAxisAngle + Math.PI / 2; // Perpendicular to text direction
-
-  // Step 1: Find boundary lines using perpendicular projection
-  // Project each line's center onto the secondary axis (perpendicular to text direction)
-  const lineProjections = lines.map((line) => {
-    const lineCenterX =
-      (line.top_left.x +
-        line.top_right.x +
-        line.bottom_left.x +
-        line.bottom_right.x) /
-      4;
-    const lineCenterY =
-      (line.top_left.y +
-        line.top_right.y +
-        line.bottom_left.y +
-        line.bottom_right.y) /
-      4;
-
-    const relX = lineCenterX - centroid.x;
-    const relY = lineCenterY - centroid.y;
-    const secondaryProjection =
-      relX * Math.cos(secondaryAxisAngle) + relY * Math.sin(secondaryAxisAngle);
-
-    return {
-      line,
-      projection: secondaryProjection,
-      centerX: lineCenterX,
-      centerY: lineCenterY,
-    };
-  });
-
-  // Sort by projection to find extremes
-  lineProjections.sort((a, b) => a.projection - b.projection);
-
-  // Get boundary lines (use top 20% for each boundary to handle noise)
-  const boundaryCount = Math.max(1, Math.ceil(lines.length * 0.2));
-  const leftBoundaryLines = lineProjections
-    .slice(0, boundaryCount)
-    .map((p) => p.line);
-  const rightBoundaryLines = lineProjections
-    .slice(-boundaryCount)
-    .map((p) => p.line);
-
-  // Step 2: Get edge points from boundary lines
-  let leftEdgePoints: Point[] = [];
-  let rightEdgePoints: Point[] = [];
-
-  // For left boundary lines, get their leftmost edges
-  leftBoundaryLines.forEach((line) => {
-    const leftX = Math.min(line.top_left.x, line.bottom_left.x);
-    leftEdgePoints.push(
-      { x: leftX, y: line.top_left.y },
-      { x: leftX, y: line.bottom_left.y }
-    );
-  });
-
-  // For right boundary lines, get their rightmost edges
-  rightBoundaryLines.forEach((line) => {
-    const rightX = Math.max(line.top_right.x, line.bottom_right.x);
-    rightEdgePoints.push(
-      { x: rightX, y: line.top_right.y },
-      { x: rightX, y: line.bottom_right.y }
-    );
-  });
-
-  return {
-    leftEdge: leftEdgePoints,
-    rightEdge: rightEdgePoints,
-  };
-};
 
 /**
  * Find the top and bottom edges of lines at the secondary axis extremes
